added shapes, color and motion + rotation
I added in a torus and cylinders to the original project and turned the cylinders into cones. I attempted to make the cones bend but was unable to achieve this.

I then added more boxs and gave them different rotations, movements and speeds.

Finally I changed the color.... this last one will be a challenge for me to learn.